Add full project with code, docs, and tests
