+++ example/sample_dos.py
@@ -1,7 +1,7 @@
 import random
 
-from measure_dns import DNSFlags, DNSQuery, send_dns_query
-
+from measure_dns import DNSFlags, DNSQuery
+from measure_dns import send_dns_query
 if __name__ == "__main__":
     N = 1000
 
@@ -27,7 +27,7 @@ if __name__ == "__main__":
     # Other available nameservers (commented out)
     # dns_server = "2406:da1a:8e8:e8cb:97fe:3833:8668:54ad" # ns2.testprotocol.in
     # dns_server = "2406:da18:c78:2b8:a93c:708c:4fc7:f75d" # ns3.testprotocol.in
-    dns_server = "2406:da18:c78:219:22a5:8271:5f0d:780b"  # ns4.testprotocol.in
+    dns_server = "13.127.175.92"  # ns4.testprotocol.in
     # dns_server = "65.0.92.216" # ns2.testprotocol.in (IPv4)
     # dns_server = "13.127.175.92" # ns1.testprotocol.in (IPv4)
     query_list = []
