+++ requirements.txt
@@ -1,2 +1,4 @@
 wheel
 dnspython
+pytest
+mkdocs
